piseo_etuve/requirements.txt
python-dotenv
matplotlib
Django
pyModbusTCP
pillow